// back-end/src/core/enums/scenario.enum.ts
export enum TypeWhatsappMessage {
    BUTTON_REPLY = 'button_reply',
    TEXT = 'text',
    LIST_REPLY = 'list_reply',
    INTERACTIVE = 'interactive',
    IMAGE = 'image',
    BUTTON_TEMPLATE = 'button'
}

export enum TypeSendWhatsappMessage {
    BUTTON = 'button',
    TEXT = 'text',
    LIST = 'list',
    INTERACTIVE = 'interactive',
    IMAGE = 'image',
    CATALOGUE = 'catalog_message'
}